style: redis edit and new form (#2855)
Summary:
- Unify the virtual size handling code
- To integer parsing that passes through
- new fileds on redis

Test Plan:
- Tests added

+++ platform_umbrella/apps/control_server/priv/repo/migrations/20210501181513_create_pg_clusters.exs
@@ -7,7 +7,7 @@ defmodule ControlServer.Repo.Migrations.CreateClusters do
       add :name, :string
       add :storage_size, :bigint
       add :storage_class, :string
-      add :cpu_requested, :integer
+      add :cpu_requested, :bigint
       add :cpu_limits, :integer
       add :memory_requested, :bigint
       add :memory_limits, :bigint
